+++ Civi/ActionProvider/Action/Communication/CreatePdf.php
@@ -16,12 +16,19 @@ use CRM_ActionProvider_ExtensionUtil as E;
 
 class CreatePdf extends AbstractAction {
 
+  /**
+   * @var \ZipArchive
+   */
+  protected $zip;
+
   public function doAction(ParameterBagInterface $parameters, ParameterBagInterface $output) {
     $domain     = \CRM_Core_BAO_Domain::getDomain();
     $message = $parameters->getParameter('message');
     $contactId = $parameters->getParameter('contact_id');
     $filename = $this->configuration->getParameter('filename');
-    $filename .= '_' . $contactId . '.pdf';
+    $fileNameWithoutContactId = $filename . '.pdf';
+    $filenameWithContactId = $filename . '_' . $contactId . '.pdf';
+    $subdir = $this->createSubDir($this->currentBatch);
 
     $contact = civicrm_api3('Contact', 'getsingle', array('id' => $contactId));
 
@@ -44,10 +51,10 @@ class CreatePdf extends AbstractAction {
       $message = $smarty->fetch("string:{$message}");
     }
 
-    $contents = \CRM_Utils_PDF_Utils::html2pdf($message, $filename, TRUE);
-
-    $fullFilePath = FileWriter::writeFile($contents, $filename);
-    $mimeType = mime_content_type($fullFilePath);
+    $contents = \CRM_Utils_PDF_Utils::html2pdf($message, $filenameWithContactId, TRUE);
+    if ($this->zip) {
+      $this->zip->addFromString($filenameWithContactId, $contents);
+    }
 
     $activityParams = array(
       'activity_type_id' => \CRM_Core_PseudoConstant::getKey('CRM_Activity_BAO_Activity', 'activity_type_id', 'Print PDF Letter'),
@@ -56,16 +63,77 @@ class CreatePdf extends AbstractAction {
       'target_contact_id' => $contactId,
     );
     $result = civicrm_api3('Activity', 'create', $activityParams);
-    civicrm_api3('Attachment', 'create', array(
+    $attachment = civicrm_api3('Attachment', 'create', array(
       'entity_table' => 'civicrm_activity',
       'entity_id' => $result['id'],
-      'name' => $filename,
-      'mime_type' => $mimeType,
+      'name' => $fileNameWithoutContactId,
+      'mime_type' => 'application/pdf',
       'content' => $contents,
     ));
 
-    $downloadUrl = \CRM_Utils_System::url('civicrm/actionprovider/downloadfile', array('filename' => $filename));
-    \CRM_Core_Session::setStatus(E::ts('Created document for %1 <a href="%2">Download document<a/>', array(1=>$contact['display_name'], 2=>$downloadUrl)));
+    $file = reset($attachment['values']);
+
+    $output->setParameter('filename', $file['name']);
+    $output->setParameter('url', $file['url']);
+    $output->setParameter('path', $file['path']);
+  }
+
+  /**
+   * This function initialize a batch.
+   *
+   * @param $batchName
+   */
+  public function initializeBatch($batchName) {
+    // Child classes could override this function
+    // E.g. create a directory
+    $this->createSubDir($batchName);
+
+    $subdir = $this->createSubDir();
+    $outputName = \CRM_Core_Config::singleton()->templateCompileDir . $subdir.'/'.$batchName.'.zip';
+    $this->zip = new \ZipArchive();
+    if ($this->zip->open($outputName, \ZipArchive::CREATE) !== TRUE) {
+      $this->zip = null;
+    }
+
+    $this->currentBatch = $batchName;
+  }
+
+  /**
+   * This function finishes a batch and is called when a batch with actions is finished.
+   *
+   * @param $batchName
+   * @param bool
+   *   Whether this was the last batch.
+   */
+  public function finishBatch($batchName, $isLastBatch=false) {
+    // Child classes could override this function
+    // E.g. merge files in a directorys
+    $subdir = $this->createSubDir();
+    $downloadName = $this->configuration->getParameter('filename').'.zip';
+    if ($this->zip) {
+      $this->zip->close();
+
+      if ($isLastBatch) {
+        $downloadUrl = \CRM_Utils_System::url('civicrm/actionprovider/downloadfile', [
+          'filename' => $batchName . '.zip',
+          'subdir' => $subdir,
+          'downloadname' => $downloadName
+        ]);
+        \CRM_Core_Session::setStatus(E::ts('<a href="%1">Download document(s)<a/>', [1 => $downloadUrl]), E::ts('Created PDF'), 'success');
+      }
+    }
+  }
+
+  protected function createSubDir() {
+    $subDir = 'action_provider';
+    $basePath = \CRM_Core_Config::singleton()->templateCompileDir . $subDir;
+    \CRM_Utils_File::createDir($basePath);
+    \CRM_Utils_File::restrictAccess($basePath.'/');
+    $subDir .= '/createpdf';
+    $basePath = \CRM_Core_Config::singleton()->templateCompileDir . $subDir;
+    \CRM_Utils_File::createDir($basePath);
+    \CRM_Utils_File::restrictAccess($basePath.'/');
+    return $subDir;
   }
 
   /**
@@ -83,6 +151,14 @@ class CreatePdf extends AbstractAction {
   public function getConfigurationSpecification() {
     return new SpecificationBag(array(
       new Specification('filename', 'String', E::ts('Filename'), true, E::ts('document')),
+    ));
+  }
+
+  public function getOutputSpecification() {
+    return new SpecificationBag(array(
+      new Specification('filename', 'String', E::ts('Filename')),
+      new Specification('url', 'String', E::ts('Download Url')),
+      new Specification('path', 'String', E::ts('Path in filesystem')),
     ));
   }